requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.111.0
fastapi-cli==0.0.24
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
imbalanced-learn==0.12.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.19.0
numpy==1.26.4
orjson==3.11.8
packaging==24.2
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.9
pyarrow==23.0.1
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.9.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.0
rich==13.9.4
rich-toolkit==0.19.7
rpds-py==0.30.0
s3transfer==0.9.0
scikit-learn==1.4.2
scipy==1.12.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.30
starlette==0.37.2
streamlit==1.35.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.4.1
tornado==6.5.5
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
